# internal/argo-controller/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "SPX Argo Controller HTTP API",
        "title": "SPX Argo Controller",
        "contact": {
            "name": "API Support",
            "url": "https://superphenix.net",
            "email": "contact@superphenix.net"
        },
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/{orgId}/{projectId}/mark": {
            "get": {
                "security": [
                    {
                        "Bearer": [
                            ""
                        ]
                    }
                ],
                "description": "Mark all argo resources inside a project for deletion",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1",
                    "GC"
                ],
                "summary": "Mark for deletion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "orgId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
...
    "securityDefinitions": {
        "Bearer": {
            "description": "Auth Token to have access to Argo Ctrl. \"Bearer [Auth Token]\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "scopes": {
                "organizationbillingread": "Grants read access to organization billing",
                "organizationbillingwrite": "Grants write access to organization billing",
                "organizationiamread": "Grants read access to organization IAM",
                "organizationiamwrite": "Grants write access to organization IAM",
                "organizationprojectmanagement": "Grants project management access in organization",
                "organizationread": "Grants read access to organization",
                "organizationwrite": "Grants write access to organization",
                "projectdiskread": "Grants read access to Disk in project",
                "projectdiskwrite": "Grants write access to Disk in project",
                "projecteipread": "Grants read access to EIP in project",
                "projecteipwrite": "Grants write access to EIP in project",
                "projectinstancecontrol": "Grants control (start, stop) access to instance in project",
                "projectinstanceread": "Grants read access to instance in project",
                "projectinstanceterminal": "Grants terminal access to instance in project",
                "projectinstancewrite": "Grants write access to instance in project",
                "projectlbread": "Grants read access to LoadBalancer in project",
                "projectlbwrite": "Grants write access to LoadBalancer in project",
                "projectsnapshotread": "Grants read access to Snapshot and Instance Snapshot in project",
                "projectsnapshotwrite": "Grants write access to Snapshot and Instance Snapshot in project",
                "projectsshread": "Grants read access to SSH in project",
                "projectsshwrite": "Grants write access to SSH in project",
                "projectsubnetread": "Grants read access to Subnet in project",
                "projectsubnetwrite": "Grants write access to Subnet in project",
                "projectvpcread": "Grants read access to VPC in project",
                "projectvpcwrite": "Grants write access to VPC in project"
            }
        }
    }
}